Fixed ASM code %define issues that caused segfaults
There are many places in the assembly code that expect the function's argument to be in a specific register,
but modern GCC does not follow the calling convention. This is because the assembly code expects __GCC2 to be
defined to follow the correct behavior. Putting %define __GCC2 at the top of the file or commenting out the
%ifdef __GCC2 is a bandaid fix that works, and Gens seems to run fine with this. I will try to find a better
solution for this in the future.