Fixed GCC not using fastcall for assembly functions
Previously, the code used attribute regparm(2) to specify that arguments should be passed
in through registers, but this produces incorrect code with modern GCC given the assembly code's
calling conventions. I replaced the relevant #defines with attribute fastcall, which fixed the issues.
The additions will need to be tested with other compilers.